% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsdistcomp.R
\docType{class}
\name{Computation}
\alias{Computation}
\title{Class that manages Computation in the MRSDistComp application.
It registers all existing computations in the distcomp package
available on a specific site for use in model definitions}
\format{\code{\link{R6Class}} object.}
\usage{
Computation
}
\value{
Object of \code{\link{R6Class}} with methods for communication with backend MRSDistComp database.
}
\description{
Class that manages Computation in the MRSDistComp application.
It registers all existing computations in the distcomp package
available on a specific site for use in model definitions
}
\section{Fields}{

\describe{
\item{\code{Id}}{Unique identifier for the Computation}
}}

\examples{
Computation$new()
}
\keyword{datasets}
