% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsdistcomp.R
\docType{class}
\name{ComputationInfo}
\alias{ComputationInfo}
\title{Class that defines a formula for a distributed computation "project".
This enforces two demands - that the formula fields are bound by a schema (DataCatalog) and
that the actual "model fitting" be performed using a specific type of learner
that is included in the "distcomp" package such as - Stratified Cox or RankKSVD.}
\format{\code{\link{R6Class}} object.}
\usage{
ComputationInfo
}
\value{
Object of \code{\link{R6Class}} with methods for communication with backend MRSDistComp database.
}
\description{
Class that defines a formula for a distributed computation "project".
This enforces two demands - that the formula fields are bound by a schema (DataCatalog) and
that the actual "model fitting" be performed using a specific type of learner
that is included in the "distcomp" package such as - Stratified Cox or RankKSVD.
}
\section{Fields}{

\describe{
\item{\code{Id}}{Unique identifier for a ComputationInfo object.}

\item{\code{Name}}{Name of the ComputationInfo object.}

\item{\code{Formula}}{Formula of the ComputationInfo. Expected to be R code.}

\item{\code{DataCatalog}}{Schema for the ComputationInfo. The DataCatalog is an empty column based on a sample data frame provided by the customer.}

\item{\code{ComputationType}}{Type of computation this ComputationInfo uses Must be one of the supported computations in the "distcomp" package.}

\item{\code{IsEnabled}}{Specify whether the ComputationInfo is enabled or disabled. Temporarily disable to control access and execution.}
}}

\section{Methods}{


\describe{
  \item{\code{new()}}{This method is used to create object of the ComputationInfo class.}
  \item{\code{createComputationInfo(name,formula,datacatalog,computationtype,isenabled=1,validfrom=NULL,validto=NULL)}}{This method creates a new ComputationInfo in the database. Returns Id of the newly created ComputationInfo.}
  \item{\code{getAllComputationInfo()}}{This method gets the list of all models.}
  \item{\code{getComputationInfoByName(nameFilter=NULL)}}{This method gets the list of all models by name.Empty nameFilter returns all computationinfo objects.}}
}

\examples{
ComputationInfo$new()
}
\keyword{datasets}
