% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsdistcomp.R
\docType{class}
\name{ComputationInfoJob}
\alias{ComputationInfoJob}
\title{Class that manages ComputationInfoJob in MRSDistComp application.}
\format{\code{\link{R6Class}} object.}
\usage{
ComputationInfoJob
}
\value{
Object of \code{\link{R6Class}} with methods for communication with backend MRSDistComp database.
}
\description{
Class that manages ComputationInfoJob in MRSDistComp application.
}
\section{Fields}{

\describe{
\item{\code{Id}}{Unique identifier for the Computation job}

\item{\code{ComputationInfo}}{Project Name this Computation job is for.}

\item{\code{Result}}{Result object of a computation job.}

\item{\code{Summary}}{Summary of the computation job.}

\item{\code{LogTxt}}{Log output of the computation job.}

\item{\code{Status}}{Status of the ComputationInfoJob.Expected values are - 'NotStarted','Running','Completed','CompletedWithErrors'}

\item{\code{StartDateTime}}{Start datetime of this ComputationInfoJob.}

\item{\code{EndDateTime}}{Datetime when this ComputationInfoJob ended.}
}}

\section{Methods}{


\describe{
  \item{\code{new()}}{This method is used to create object of this class.}
  \item{\code{triggerJob(projectname,status='Created',startdatetime=NULL,enddatetime=NULL)}}{This method creates a new computation job for a ComputationInfo i.e., "project".
  This method checks the list of participants for a project and if any found, sends them a new "synchronous" computation request. It is expected that
  the "sites" or "workers" already have the ComputationInfo definition (with the same name) and its dependencies - i.e., data catalog (schema)
  data sources for the schema defined. This method will also be exposed as a webservice, ONLY on the "master".}
  \item{\code{processJob(projectname,status='Created',startdatetime=NULL,enddatetime=NULL)}}{This method creates a new computation job for a ComputationInfo i.e., "project". This is exposed as a webservice only on the "master".}
  \item{\code{getJobsByStatus(statusFilter)}}{This method gets a list of all jobs based on a status. Possible status values are: 'Created', 'InProgress', 'Completed', 'Failed'}
  \item{\code{getJobsForProject(projectName)}}{This method gets a list of all computation jobs for a project.}}
}

\examples{
ComputationInfoJob$new()
}
\keyword{datasets}
