% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsdistcomp.R
\docType{class}
\name{ComputationInfoParticipants}
\alias{ComputationInfoParticipants}
\title{Relationship class that manages the participants in computation projects.
A participant can be in one or more computation projects and there can be
any number of computation projects defined in the system.}
\format{\code{\link{R6Class}} object.}
\usage{
ComputationInfoParticipants
}
\value{
Object of \code{\link{R6Class}} with methods for communication with backend MRSDistComp database.
}
\description{
Relationship class that manages the participants in computation projects.
A participant can be in one or more computation projects and there can be
any number of computation projects defined in the system.
}
\section{Fields}{

\describe{
\item{\code{Id}}{Unique identifier for a ComputationInfoParticipant object}

\item{\code{ComputationInfoName}}{Name of the Computation project}

\item{\code{ParticipantName}}{Name of the participant or site.}

\item{\code{IsEnabled}}{Specify whether the ComputationInfoParticipant is enabled or disabled. Use this flag to control who participates in a specific Computation project.}
}}

\section{Methods}{


\describe{
  \item{\code{new()}}{This method is used to create object of the Model class.}
  \item{\code{create(id,computationinfoname,participantname,isenabled=1)}}{This method creates a new ModelParticipant that links a participant or Site with a Model.}
  \item{\code{getAllComputationInfoParticipants()}}{This method gets the list of all ComputationInfoParticipants records.}
  \item{\code{getParticipantsOfAProject(computationInfoNameFilter)}}{This method returns a data.frame that contains all participants for a specific computation project}}
  \item{\code{getProjectsOfAParticipant(participantNameFilter)}}{This method returns a data.frame that contains all computation projects a specific participant is registered for.}
}

\examples{
ComputationInfoParticipants$new()
}
\keyword{datasets}
