% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsdistcomp.R
\docType{class}
\name{DataCatalog}
\alias{DataCatalog}
\title{Class that manages data catalog or entities
in the MRSDistComp application.}
\format{\code{\link{R6Class}} object.}
\usage{
DataCatalog
}
\value{
Object of \code{\link{R6Class}} with methods for communication with backend MRSDistComp database.
}
\description{
Class that manages data catalog or entities
in the MRSDistComp application.
}
\section{Fields}{

\describe{
\item{\code{Id}}{Unique identifier for the DataCatalog}

\item{\code{Name}}{Unique name for a schema that will be used in a computation}

\item{\code{Description}}{Description for the schema}

\item{\code{Version}}{Version of the Schema. Name is currently the only primary key. Version can be set to any value.}

\item{\code{Schema}}{Actual schema content. This is currently a serialized 'exemplar' data frame (i.e., an empty data frame).}
}}

\examples{
DataCatalog$new()
}
\keyword{datasets}
