% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsdistcomp.R
\docType{class}
\name{DataSources}
\alias{DataSources}
\title{Class that manages data sources for data to be used in the DistComp application.}
\format{\code{\link{R6Class}} object.}
\usage{
DataSources
}
\value{
Object of \code{\link{R6Class}} with methods for communication with backend MRSDistComp database.
}
\description{
Class that manages data sources for data to be used in the DistComp application.
}
\section{Fields}{

\describe{
\item{\code{Id}}{Unique identifier for a ComputationInfoParticipant object}

\item{\code{Name}}{Name of a Data Source}

\item{\code{Description}}{Description for the Data Source}

\item{\code{Type}}{Type of Data Source. Any meaningful identifier such as "CSV", "SQLAzure", "AzureBlob", "GoogleBlob" etc.,}

\item{\code{DataCatalog}}{Name of the DataCatalog (Schema) that this data source conforms to. Both to keep track of from the end-user as well as for the distcomp application.}

\item{\code{AccessInfo}}{Content that contains access related information for the data source. For e.g., a connection string to a database}
}}

\section{Methods}{


\describe{
  \item{\code{new()}}{This method is used to create a new DataSources class object.}
  \item{\code{createDataSource(name,description,datasourcetype,datacatalogname,accessinfo)}}{This method creates a new DataSource for use with DistComp application.No restrictions are enforced as to what the source is expected to be.}
  \item{\code{getDataSourcesForSchema(schemaNameFilter)}}{This method gets all data sources defined for a specific schema.}
  \item{\code{getDataSourcesByName(nameFilter)}}{This method returns a data.frame for a data sources with the given name. Specifying NULL value for the nameFilter returns all datasource records defined.}}
}

\examples{
DataInfo$new()
}
\keyword{datasets}
