% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlsdistcomp.R
\docType{class}
\name{Participant}
\alias{Participant}
\title{Class that manages all Participants or collaborators in
the MRSDistComp application.}
\format{\code{\link{R6Class}} object.}
\usage{
Participant
}
\value{
Object of \code{\link{R6Class}} with methods for communication with backend MRSDistComp database.
}
\description{
Class that manages all Participants or collaborators in
the MRSDistComp application.
}
\section{Fields}{

\describe{
\item{\code{Id}}{Unique identifier for the participant.}

\item{\code{Name}}{Name of the participant.}

\item{\code{ClientId}}{Application Id of the participant's Azure Active Directory Application.}

\item{\code{TenantId}}{Azure Active Directory Tenant name of the participant.}

\item{\code{ClientSecret}}{Client secret of the participant's Azure Active Directory Application.}

\item{\code{Url}}{Base Url of the MRS deployment. FOr e.g., https://mrsdistcomp1.westus.app.microsoft.com:12800/api.}

\item{\code{IsEnabled}}{Specifies if the participant is valid or enabled to be able to perform any operations.}

\item{\code{ValidFrom}}{Date since when the participant is active.}

\item{\code{ValidTo}}{Date till when the participant is active.Past date invalidates participant.}
}}

\examples{
Participant$new()
}
\keyword{datasets}
